number of subarrys with xor k

+++ array/4sum.cpp
@@ -6,7 +6,7 @@ using namespace std;
 
 int main(){
     vector<int>nums{1,1,1,2,2,2,3,3,3,4,4,4,5,5};
-    int target;
+    long long target;
     cout<<"Enter target: ";
     cin>>target;
     int n = nums.size();
